Add util for export file name
