+++ src/app/utils/constants.ts
@@ -199,4 +199,6 @@ export const DEFAULT_LAST_LOGIN_DATE = '0001-01-01T00:00:00';
 
 export const DEBOUNCE_TIME = 500;
 
-export const UserExportFileName = 'Users.xlsx';
+export const USER_EXPORT_FILE_PREFIX = 'User';
+
+export const EXPORT_FILE_NAME_TEMPLATE = '{prefix}_export_{date}.xlsx';
